services/control-plane/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pyyaml==6.0.2
networkx==3.3
prometheus-client==0.20.0
